+++ Developpement/interactify_app/lib/l10n/app_fr.arb
@@ -1,11 +1,18 @@
 {
-    "appTitle": "flutter_skeleton_1",
-    "@appTitle": {
-      "description": "Le titre de l'application"
-    },
-    "appName": "Mon app",
-    "@appName": {
-      "description": "Le nom de l'application"
-    }
+  "appTitle": "Interactify_Application",
+  "@appTitle": {
+    "description": "Le titre de l'application"
+  },
+  "appName": "Interactify_Application",
+  "@appName": {
+    "description": "Le nom de l'application"
+  },
+  "appTest": "'Appuyez sur le boutton:',",
+  "@appTest": {
+    "description": "boutton texte"
+  },
+  "appFull": "full",
+  "@appFull": {
+    "description": "full"
   }
-  
+}
